should use double

+++ le-scoring/src/main/java/com/latticeengines/scoring/service/impl/ScoringManagerServiceImpl.java
@@ -104,7 +104,7 @@ public class ScoringManagerServiceImpl extends QuartzJobBean implements ScoringM
 
     @VisibleForTesting
     void cleanTables() {
-        int cleanUpInterval = Integer.parseInt(this.cleanUpInterval);
+        double cleanUpInterval = Double.parseDouble(this.cleanUpInterval);
 
          List<ScoringCommand> consumedCommands = scoringCommandEntityMgr.getConsumed();
          for(ScoringCommand scoringCommand : consumedCommands){
